+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/controller/PostController.java
@@ -19,8 +19,6 @@ import wooteco.team.ittabi.legenoaroundhere.dto.PageRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.PageableAssembler;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostCreateRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostImageResponse;
-import wooteco.team.ittabi.legenoaroundhere.dto.PostReportCreateRequest;
-import wooteco.team.ittabi.legenoaroundhere.dto.PostReportResponse;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostResponse;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostSearchRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostUpdateRequest;
@@ -97,16 +95,6 @@ public class PostController {
 
         return ResponseEntity
             .noContent()
-            .build();
-    }
-
-    @PostMapping("/{postId}/reports")
-    public ResponseEntity<Void> createPostReport(@PathVariable Long postId, @RequestBody
-        PostReportCreateRequest postReportCreateRequest) {
-        final PostReportResponse postReportResponse = postService
-            .createPostReport(postId, postReportCreateRequest);
-        return ResponseEntity
-            .created(URI.create("/posts/" + postReportResponse.getId()))
             .build();
     }
 }
