+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/service/ReportService.java
@@ -5,16 +5,41 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.team.ittabi.legenoaroundhere.config.IAuthenticationFacade;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.Post;
 import wooteco.team.ittabi.legenoaroundhere.domain.post.PostReport;
+import wooteco.team.ittabi.legenoaroundhere.domain.user.User;
+import wooteco.team.ittabi.legenoaroundhere.dto.PostReportAssembler;
+import wooteco.team.ittabi.legenoaroundhere.dto.PostReportCreateRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostReportResponse;
 import wooteco.team.ittabi.legenoaroundhere.exception.NotExistsException;
 import wooteco.team.ittabi.legenoaroundhere.repository.PostReportRepository;
+import wooteco.team.ittabi.legenoaroundhere.repository.PostRepository;
 
 @Service
 @AllArgsConstructor
 public class ReportService {
 
     private final PostReportRepository postReportRepository;
+    private final PostRepository postRepository;
+    private final IAuthenticationFacade authenticationFacade;
+
+    @Transactional
+    public PostReportResponse createPostReport(Long postId,
+        PostReportCreateRequest postReportCreateRequest) {
+        User user = (User) authenticationFacade.getPrincipal();
+        Post post = findPostBy(postId);
+
+        PostReport postReport = PostReportAssembler.assemble(postReportCreateRequest, post, user);
+
+        PostReport savedPostReport = postReportRepository.save(postReport);
+        return PostReportResponse.of(savedPostReport);
+    }
+
+    private Post findPostBy(Long id) {
+        return postRepository.findById(id)
+            .orElseThrow(() -> new NotExistsException("ID에 해당하는 POST가 없습니다."));
+    }
 
     @Transactional(readOnly = true)
     public PostReportResponse findPostReport(Long id) {
